Rename: 레이아웃 관련 파일들 src/layout으로 이동, src/utils/datas 정의
