--- src/components/Home.jsx
@@ -1,5 +1,0 @@
-const Home = () => {
-  return <div>Home</div>;
-};
-
-export default Home;
