+++ src/layout/PageLayout.jsx
@@ -1,5 +1,5 @@
-import { Header } from "../components/Header";
-import { Footer } from "../components/Footer";
+import { Header } from "./Header";
+import { Footer } from "./Footer";
 
 const PageLayout = ({ page }) => {
   return (
