--- src/utils/utils.js
@@ -1,8 +1,0 @@
-export const categories = {
-  공부: ["어학", "수험", "프로그래밍", "기타"],
-  생활습관: ["운동", "자기계발", "식습관", "기타"],
-  자격증: ["기사", "NCS", "컴활", "한국사", "기타"],
-  취미: ["독서", "미술", "음악", "기타"],
-  경제: ["절약", "투자", "기타"],
-  기타: [],
-};
